DAT-710 Improve functionality of EntitytTagRelationship objects

+++ fd-engine/src/main/java/org/flockdata/engine/admin/service/AdminService.java
@@ -142,7 +142,7 @@ public class AdminService implements EngineAdminService {
             if ( !segment.isDefault())
                 conceptService.delete(documentType,segment);
         }
-        if ( segmentToDelete== null  )
+        if ( segmentToDelete== null && total ==0 ) // First purge is data only. Delete again to remove the docType
             // Removing the DocumentType
             conceptService.delete(documentType);
